requirements.txt
aiogram==3.11.0
google_api_python_client==2.140.0
protobuf==5.27.3
python-dotenv==1.0.1